refactor: 💡 Move client daemon handler to API addon
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-11691
