+++ addons/api/app/handlers/client-daemon-handler.js
@@ -1,0 +1,6 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+export { default } from 'api/handlers/client-daemon-handler';
